[require-refactoring] several useful refactorings: 1. Only have 'Solr' module, because solrb is just a gem name, it does not mandate the matching module.    We should only have a single module per gem. 2. Move inner requires into respective nested filed, should be cleaner
