+++ lib/solr/document.rb
