+++ lib/solr/document_collection.rb
