+++ lib/solr/grouped_document_collection.rb
