+++ lib/solr/request/boost_magnitude.rb
