+++ lib/solr/request/boosting/dictionary_boost_function.rb
@@ -21,7 +21,7 @@ module Solr
           sf = solarize_field(field)
           dictionary.to_a.reverse.reduce(1) do |acc, (field_value, boost)|
             if field_value.is_a?(String) || field_value.is_a?(Symbol)
-              "if(termfreq(#{sf},\"#{Solrb::Utils.solr_escape(field_value.to_s)}\"),#{boost},#{acc})"
+              "if(termfreq(#{sf},\"#{Solr::Utils.solr_escape(field_value.to_s)}\"),#{boost},#{acc})"
             else
               "if(eq(#{sf},#{field_value}),#{boost},#{acc})"
             end
