+++ lib/solr/request/boosting/exists_boost_function.rb
