+++ lib/solr/request/boosting/field_value_less_than_boost_function.rb
