+++ lib/solr/request/boosting/field_value_match_boost_function.rb
