+++ lib/solr/request/boosting/geodist_function.rb
