+++ lib/solr/request/boosting/ln_function_boost.rb
