+++ lib/solr/request/boosting/numeric_field_value_match_boost_function.rb
