+++ lib/solr/request/boosting/phrase_proximity_boost.rb
