+++ lib/solr/request/boosting/ranking_field_boost_function.rb
