+++ lib/solr/request/boosting/recent_field_value_boost_function.rb
