+++ lib/solr/request/boosting/scale_function_boost.rb
