+++ lib/solr/request/boosting/textual_field_value_match_boost_function.rb
