+++ lib/solr/request/edismax_adapter.rb
