+++ lib/solr/request/facet.rb
