+++ lib/solr/request/field_with_boost.rb
