+++ lib/solr/request/geo_filter.rb
