+++ lib/solr/request/grouping.rb
