+++ lib/solr/request/limit_docs_by_field.rb
