+++ lib/solr/request/or_filter.rb
