+++ lib/solr/request/sorting.rb
