+++ lib/solr/request/sorting/field.rb
