+++ lib/solr/request/spellcheck.rb
