+++ lib/solr/response/edismax_adapter.rb
