+++ lib/solr/response/facet_value.rb
