+++ lib/solr/response/field_facets.rb
