+++ lib/solr/version.rb
@@ -1,3 +1,3 @@
-module Solrb
+module Solr
   VERSION = "0.1.0"
 end
