+++ solrb.gemspec
@@ -1,16 +1,15 @@
-
 lib = File.expand_path("../lib", __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require "solrb/version"
+require "solr/version"
 
 Gem::Specification.new do |spec|
   spec.name          = "solrb"
-  spec.version       = Solrb::VERSION
-  spec.authors       = ["Adriano Luz"]
-  spec.email         = ["adriano.luz@machinio.com"]
+  spec.version       = Solr::VERSION
+  spec.authors       = ["Adriano Luz", "Valentin Vasilyev"]
+  spec.email         = ["adriano.luz@machinio.com", "valentin@machinio.com"]
 
-  spec.summary       = "Solr Ruby client"
-  spec.homepage      = "https://github.com/machinio/solrb"
+  spec.summary       = "Solr Ruby client with a nice object-oriented API"
+  spec.homepage      = "https://github.com/machinio/solr"
   spec.license       = "MIT"
 
   # Prevent pushing this gem to RubyGems.org. To allow pushes either set the 'allowed_push_host'
